demo v3 Remove item was addedo to Home in deleteButton

+++ src/utilities/Menu.java
@@ -17,7 +17,11 @@ public class Menu {
         RepositoryProductsImpl.addProduct(new Product(name, price, stock));
     }
     public static void deleteProduct(int id){
-        RepositoryProductsImpl.deleteProduct(id);
+        if (id <= RepositoryProductsImpl.sizeProducts()){
+            RepositoryProductsImpl.deleteProduct(id);
+        }else {
+
+        }
     }
 
 }
